Forgot two Integers - now it should be fine
Please double check Stefan

+++ database/tables/scan_details.sql
@@ -2,5 +2,5 @@ CREATE TABLE scan_details (
     id BIGSERIAL PRIMARY KEY,
     scan_id bigint NOT NULL REFERENCES scans(id),
     file_id bigint NOT NULL REFERENCES files(id),
-    checksum_id INTEGER NOT NULL REFERENCES checksums(id)
+    checksum_id bigint NOT NULL REFERENCES checksums(id)
 );
